a11

+++ cspp2/cspp2-practice/m7/Student.java
@@ -9,7 +9,7 @@ public class Student {
     */
     String name, rollnumber;
     /**
-     * to calculate average
+     * to calculate average.
      *
      * @param      a     marks.
      * @param      b     marks.
@@ -19,7 +19,7 @@ public class Student {
         /**
          * gpa calculation.
          */
-        System.out.println((double)((int)(((a + b + c) / 3)*10.0))/10.0);
+        System.out.println(((int)(((a + b + c) / 3)*10.0))/10.0);
     }
     Student(String name, String rollnumber, double a, double b, double c) {
         //function
